sangharsh-project/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-ckeditor==6.4.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
